pyproject.toml
[project]
name = "dwarf-bench"
version = "0.1.0"
description = "Benchmark that tests LLM offline knowledge of obscure dwarf facts across fantasy media."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
dwarf-bench = "dwarf_bench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dwarf_bench"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
